separate pulsar into its own file
